加上心率计算

+++ app/src/main/java/com/example/mrchenrunfeng/myecg/classes/Command.java
@@ -57,7 +57,7 @@ public class Command {
     //数据暂存队列 阻塞队列
     public static LinkedBlockingQueue<Short> mShowDataQueue=new LinkedBlockingQueue<Short>() ;
     public static Queue<Short> mAboutheartratedataListQueue=new LinkedBlockingQueue<Short>();
-    public static Queue<Byte> mHeartRateQueue=new LinkedBlockingQueue<Byte>();
+    public static Queue<Long> mHeartRateQueue=new LinkedBlockingQueue<Long>();
     public static final short ESCAPE_CHAR = 0xdb;
     public static final short SPECIAL_CHAR = 0xdd;
     public static final short SPECIAL_CHAR1 = 0xda;
